refactor: moved user request and response to folder request and responses
