+++ src/main/java/com/tcc5/car_price_compare/controllers/FavoriteController.java
@@ -1,7 +1,7 @@
 package com.tcc5.car_price_compare.controllers;
 
-import com.tcc5.car_price_compare.domain.user.dto.FavoriteRequestDTO;
-import com.tcc5.car_price_compare.domain.user.dto.FavoriteResponseDTO;
+import com.tcc5.car_price_compare.domain.request.user.FavoriteRequestDTO;
+import com.tcc5.car_price_compare.domain.response.user.FavoriteResponseDTO;
 import com.tcc5.car_price_compare.domain.user.features.Favorite;
 import com.tcc5.car_price_compare.services.ConversionService;
 import com.tcc5.car_price_compare.services.FavoriteService;
