+++ src/main/java/com/tcc5/car_price_compare/controllers/NotificationController.java
@@ -1,7 +1,7 @@
 package com.tcc5.car_price_compare.controllers;
 
-import com.tcc5.car_price_compare.domain.user.dto.NotificationRequestDTO;
-import com.tcc5.car_price_compare.domain.user.dto.NotificationResponseDTO;
+import com.tcc5.car_price_compare.domain.request.user.NotificationRequestDTO;
+import com.tcc5.car_price_compare.domain.response.user.NotificationResponseDTO;
 import com.tcc5.car_price_compare.domain.user.features.Notification;
 import com.tcc5.car_price_compare.services.ConversionService;
 import com.tcc5.car_price_compare.services.NotificationService;
