+++ src/main/java/com/tcc5/car_price_compare/domain/response/user/FavoriteResponseDTO.java
@@ -1,4 +1,4 @@
-package com.tcc5.car_price_compare.domain.user.dto;
+package com.tcc5.car_price_compare.domain.response.user;
 
 import com.tcc5.car_price_compare.domain.vehicle.dto.VehicleDTO;
 import lombok.*;
