+++ src/main/java/com/tcc5/car_price_compare/domain/user/converters/FavoriteToFavoriteResponseDTO.java
@@ -1,6 +1,6 @@
 package com.tcc5.car_price_compare.domain.user.converters;
 
-import com.tcc5.car_price_compare.domain.user.dto.FavoriteResponseDTO;
+import com.tcc5.car_price_compare.domain.response.user.FavoriteResponseDTO;
 import com.tcc5.car_price_compare.domain.user.features.Favorite;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
