+++ src/main/java/com/tcc5/car_price_compare/domain/user/converters/NotificationToNotificationResponseDTOConverter.java
@@ -1,6 +1,6 @@
 package com.tcc5.car_price_compare.domain.user.converters;
 
-import com.tcc5.car_price_compare.domain.user.dto.NotificationResponseDTO;
+import com.tcc5.car_price_compare.domain.response.user.NotificationResponseDTO;
 import com.tcc5.car_price_compare.domain.user.features.Notification;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
